feat(web): support rich text paste from Clipboard using HTML
