+++ lib/src/controller/web/quill_controller_web_real.dart
@@ -1,0 +1,34 @@
+// This file should not be exported as the APIs in it are meant for internal usage only
+
+import 'dart:async' show StreamSubscription;
+
+import 'package:web/web.dart';
+
+import '../quill_controller.dart';
+import '../quill_controller_rich_paste.dart';
+
+/// Paste event for the web.
+///
+/// Will be `null` for non-web platforms.
+///
+/// See: https://developer.mozilla.org/en-US/docs/Web/API/Element/paste_event
+StreamSubscription? _webPasteEventSubscription;
+
+extension QuillControllerWeb on QuillController {
+  void initializeWebPasteEvent() {
+    _webPasteEventSubscription =
+        EventStreamProviders.pasteEvent.forTarget(window.document).listen((e) {
+      // TODO: See if we can support markdown paste
+      final html = e.clipboardData?.getData('text/html');
+      if (html == null) {
+        return;
+      }
+      pasteHTML(html: html);
+    });
+  }
+
+  void closeWebPasteEvent() {
+    _webPasteEventSubscription?.cancel();
+    _webPasteEventSubscription = null;
+  }
+}
